casting pointers to integars

+++ 0x05-pointers_arrays_strings/1-swap.c
@@ -8,6 +8,6 @@
 
 void swap_int(int *a, int *b)
 {
-	*a = b;
-	*b = a;
+	*a = (int) b;
+	*b = (int) a;
 }
